feat(TCOMP-2258): refactor tests
